chore: restructure repository
