--- omni/module.py
@@ -1,16 +1,0 @@
-"""Code related to module handling"""
-from pathlib import Path
-from typing import Optional
-
-
-def fetch_module(module_url: str, commit: str, dest: Optional[Path] = None):
-    pass
-
-
-def execute_module():
-    # error code system: timeout, dependency missing, empty outputs
-    pass
-
-
-def validate_outputs():
-    pass
