+++ omni/workflow/formatter.py
@@ -1,0 +1,1 @@
+"""Placeholder for workflow formatter structure"""
